+++ src/components/file/Info.vue
@@ -63,7 +63,9 @@ export default {
         return ''
       }
 
-      const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB']
+      const sizes = [
+        'Bytes', 'KiB', 'MiB', 'GiB', 'TiB', 'PiB', 'EiB', 'ZiB', 'YiB'
+      ]
       const i = Math.floor(Math.log(bytes) / Math.log(1024))
 
       return parseFloat(