+++ src/components/home/Search.vue
@@ -6,11 +6,11 @@
       <div class="column is-9">
         <div class="field">
           <div class="control">
-            <tag-input
-              :tag.sync="tag"
-              :hasCompletedTag.sync="hasCompletedTag"
+            <search-input
+              :search.sync="search"
+              :hasCompletedSearch.sync="hasCompletedSearch"
               size="is-medium"
-              placeholder="search for files by tag…" />
+              placeholder="search for files by tag or constraint…" />
           </div>
         </div>
       </div>
@@ -41,14 +41,14 @@ import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import api from '@/api'
 import queryFormatter from '@/util/query-formatter'
 
-import TagInput from '@/components/general/TagInput'
+import SearchInput from '@/components/general/SearchInput'
 
 export default {
   name: 'Search',
   data: function () {
     return {
-      tag: '',
-      hasCompletedTag: false
+      search: '',
+      hasCompletedSearch: false
     }
   },
   methods: {
@@ -58,21 +58,21 @@ export default {
       this.$router.push({
         path: '/files',
         query: queryFormatter.generateDefaultFilesQuery(
-          this.tag.trim().toLowerCase()
+          this.search.trim().toLowerCase()
         )
       })
     }
   },
   watch: {
-    hasCompletedTag: function (hasCompletedTag) {
-      if (hasCompletedTag) {
+    hasCompletedSearch: function (hasCompletedSearch) {
+      if (hasCompletedSearch) {
         this.handleSubmit()
       }
     }
   },
   components: {
     FontAwesomeIcon,
-    TagInput
+    SearchInput
   }
 }
 </script>