+++ src/config/index.js
@@ -36,5 +36,20 @@ export default {
   defaultNamespaceColors: generateNamespaceColors(
     process.env.VUE_APP_DEFAULT_NAMESPACE_COLORS
   ),
-  fallbackTagColor: process.env.VUE_APP_FALLBACK_TAG_COLOR || '#0088fb'
+  fallbackTagColor: process.env.VUE_APP_FALLBACK_TAG_COLOR || '#0088fb',
+  mimeTypes: {
+    1: 'image/jpeg',
+    2: 'image/png',
+    3: 'image/gif',
+    4: 'image/bmp',
+    9: 'video/x-flv',
+    14: 'video/mp4',
+    18: 'video/x-ms-wmv',
+    20: 'video/x-matroska',
+    21: 'video/webm',
+    23: 'image/apng',
+    25: 'video/mpeg',
+    26: 'video/quicktime',
+    27: 'video/x-msvideo'
+  }
 }