+++ src/main-electron.js
@@ -24,6 +24,7 @@ import {
   faSignInAlt,
   faSignOutAlt,
   faSpinner,
+  faTools,
   faTrash,
   faUser,
   faUserPlus
@@ -104,6 +105,7 @@ library.add(
   faSignInAlt,
   faSignOutAlt,
   faSpinner,
+  faTools,
   faTrash,
   faUser,
   faUserPlus